Runtime: separate blob host functions and tests into a separate mod.
